requirements_dev.txt
pytest
coverage
Pillow
pandas
matplotlib!=3.5.2
soundfile; sys_platform != 'darwin' or (sys_platform == 'darwin' and platform.machine != 'arm64')
git+https://github.com/bastibe/python-soundfile@0.11.0b3#egg=soundfile; sys_platform == 'darwin' and platform.machine == 'arm64'
boto3
google-cloud-storage
google-cloud-aiplatform
kubernetes
moviepy
imageio
ipython
ipykernel
nbclient
scikit-learn
tensorflow>=1.15.2; sys_platform != 'darwin'
tensorflow>=1.15.2; python_version > '3.6' and sys_platform == 'darwin' and platform.machine != 'arm64'
tensorflow-macos; python_version > '3.6' and python_version < '3.10' and sys_platform == 'darwin' and platform.machine == 'arm64'
torch
torchvision
plotly
bokeh
tqdm
docker
stable_baselines3
tensorboard
gym
jax[cpu]; sys_platform == 'darwin' or sys_platform == 'linux'
fastcore; python_version > '3.6'
fastcore==1.3.29; python_version == '3.6'
pyarrow
metaflow>=2.3.5
xgboost
lightgbm
rdkit-pypi; sys_platform != 'darwin' or (sys_platform == 'darwin' and platform.machine != 'arm64')
rdkit-pypi; python_version > '3.7' and sys_platform == 'darwin' and platform.machine == 'arm64'
.[launch]
.[sweeps]; sys_platform != 'darwin' or (sys_platform == 'darwin' and platform.machine != 'arm64')
.[azure]
cloudpickle
responses
kfp
